Add initial API implementation
Rock-paper-scissor games are now possible via the api.
Using Jersey as the RESTful Web Services framework.
Jackson is used for POJO to JSON mapping.
Jersey client used for testing the Jersey server resource.
Added dependency to commons-lang3 since it is nice to have.
Game.isDraw() was added to simplify identification of draw games.