Allow multiple games and RESTful-ify API
Initial implementation wasn't really RESTful. This is closer to what
would be considered a REST API.
Allowed for multiple games to be played/stored at once.
Extracted a PlayerBet class to hold player name and bet information.
Added a Game.Status enum. This holds information about the current
status of the game and can identify if the game ended in a win or draw.
Stored the winning PlayerBet in its own field in Game (winner).
Stored bets in a single list instead of two separate fields.